+++ spec/arduino_backend_spec.rb
@@ -1,10 +1,8 @@
 require "spec_helper"
-require 'pathname'
 
 def get_sketch(dir, file)
-  File.join(File.dirname(__FILE__), dir, file)
+  Pathname.new(__FILE__).parent + dir + file
 end
-
 
 RSpec.describe ArduinoCI::ArduinoBackend do
   next if skip_ruby_tests
@@ -102,6 +100,60 @@ RSpec.describe ArduinoCI::ArduinoBackend do
 
     it "Passes a simple INO sketch at #{sketch_path_ino}" do
       expect(backend.compile_sketch(sketch_path_ino, "arduino:avr:uno")).to be true
+    end
+
+    it "Detects the bytes usage after compiling a sketch" do
+      expect(backend.compile_sketch(sketch_path_ino, "arduino:avr:uno")).to be true
+      the_bytes = backend.last_bytes_usage
+      expect(the_bytes[:globals]).to eq 9
+      expect(the_bytes[:free]).to eq 2039
+      expect(the_bytes[:max]).to eq 2048
+    end
+  end
+
+  context "--dry-run flags" do
+
+    sketch_path_ino = get_sketch("FakeSketch", "FakeSketch.ino")
+
+    before { allow(backend).to receive(:run_and_capture).and_call_original }
+
+    it "Uses --dry-run flag for arduino-cli version < 0.14.0" do
+      parsed_stdout = JSON.parse('{ "VersionString": "0.13.6" }')
+      cli_version_output = {
+        json: parsed_stdout
+      }
+      allow(backend).to receive(:capture_json).and_return cli_version_output
+
+      backend.compile_sketch(sketch_path_ino, "arduino:avr:uno")
+
+      expect(backend).to have_received(:run_and_capture).with(
+        "compile",
+        "--fqbn",
+        "arduino:avr:uno",
+        "--warnings",
+        "all",
+        "--dry-run",
+        sketch_path_ino.to_s
+      )
+    end
+
+    it "Does not use --dry-run flag for arduino-cli version >= 0.14.0" do
+      parsed_stdout = JSON.parse('{ "VersionString": "0.14.0" }')
+      cli_version_output = {
+        json: parsed_stdout
+      }
+      allow(backend).to receive(:capture_json).and_return cli_version_output
+
+      backend.compile_sketch(sketch_path_ino, "arduino:avr:uno")
+
+      expect(backend).to have_received(:run_and_capture).with(
+        "compile",
+        "--fqbn",
+        "arduino:avr:uno",
+        "--warnings",
+        "all",
+        sketch_path_ino.to_s
+      )
     end
   end
 end